Add files via upload
Recommendation systems are one of the essential applications of artificial intelligence, used to provide personalized suggestions to users based on their preferences and past behaviors. This paper focuses on a movie recommendation system that relies on data from the famous IMDb movie database while incorporating demographic information such as age, gender, and viewing behaviors. The aim is to present a mathematical and systematic approach to building an effective recommendation system.